util: Log to stderr by default
Currently, if someone calls ViceLog() or similar before the logging
subsystem is setup, the messages are lost. No code currently does this
(and if it did, it would be a bug), but the potential for lost log
messages is not ideal.

To avoid this, change serverLog.c to log such messages to stderr if
logging hasn't been setup, so messages are not lost, and to make it
easier for command-line utilities to show information logged via
ViceLog/FSLog.

With this commit, now any code in a library or subsystem that logs
something via ViceLog(0) will actually show up in command-line
utilities. No such subsystem that's actually used by any command-line
utilities makes use of ViceLog, but this may change in future commits.

Change-Id: I0df4148d8bd450f5efd1ec7d8dd85f5aaececd1a
Reviewed-on: https://gerrit.openafs.org/14025
Reviewed-by: Cheyenne Wills <cwills@sinenomine.net>
Reviewed-by: Kailas Zadbuke <kailashsz@in.ibm.com>
Tested-by: BuildBot <buildbot@rampaginggeek.com>
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>